+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `ydb.WithPreferredNodeID(ctx, nodeID)` context modifier for trying to execute queries on given nodeID
+
 ## v3.90.2
 * Set the `pick_first` balancer for short-lived grpc connection inside ydb cluster discovery attempt